refactor(config): split config.go (2241 → 12 files) by concern
Second worst-first file-size refactor in the queue. Each concern now has its
own ≤370-line same-package file. Function bodies are unchanged — pure
relocation plus minor lint-driven cleanup.

  config.go                          154   Config + lock/Clone + ErrInsecureCredentials
  config_constants.go                247   port/timing/threshold/logging constants
  config_types_pipeline.go           150   Pipeline* + adapter getters
  config_types_network.go            215   Server/ACME/Interface/VLAN/IP/Discovery
  config_types_health.go             287   HealthChecks + endpoints + thresholds
  config_types_security.go           114   Auth/SSO/Security/DHCP/Vuln/SNMP
  config_types_misc.go               110   Link/CableTest/Database/DNS/FAB/etc
  config_defaults.go                 369   DefaultConfig + default*Config builders
  config_load.go                     203   Load/Save/Migration/EnsureConfig
  config_validate.go                 256   Validate + per-section validators
  config_interface.go                157   credential updates + active iface
  config_profile.go                   86   ProfileExportFields + JSON I/O

All under the 600-line warn budget.

Drive-by cleanups forced by --new-from-rev lint on the relocated lines:

- Extracted logLevelInfo / logFormatText / logFormatJSON constants in
  config_constants.go (used by both DefaultConfig and validateLoggingConfig)
  to silence goconst warnings the literals trip in their new home.
- Reformatted two over-long struct-literal lines in defaultNetworkDiscovery
  Config and defaultAuthConfig to fit golines' 120-column budget.

Build, vet, lint (--new-from-rev=HEAD~1: 0 issues), and config tests pass.
